+++ HISTORY.md
@@ -3,6 +3,9 @@
 ## Packaging Notes
 * Fixed release builder ordering issue which led to CRLF line endings in 0.9.1 source distribution.
 
+## API Changes
+* Replace `Array.timestamp` with `Array.timestamp_range` [#616](https://github.com/TileDB-Inc/TileDB-Py/pull/616) 
+
 ## Improvements
 * Set `ArraySchema.tile_order=None` for Hilbert-ordered arrays [#609](https://github.com/TileDB-Inc/TileDB-Py/pull/609)
 * Use CIBW to build release wheels on Linux [#613](https://github.com/TileDB-Inc/TileDB-Py/pull/613)